multiAgent/requirements.txt
langchain
langgraph
langgraph-checkpoint
langchain-openai
python-dotenv
pydantic
langchain_text_splitters
langchain_community
langchain-tavily
feedparser
python-dateutil
beautifulsoup4
requests